fix: input auto focus 기능 수정
- backspace 눌렀을 경우 이전 input으로 이동할 수 있게 수정
