+++ src/constants.js
@@ -13,4 +13,5 @@ export const CARD_INFO_RULES = {
   HOLDER_NAME_MAX_LENGTH: 30,
   SECURITY_CODE_LENGTH: 3,
   PASSWORD_LENGTH: 2,
+  PASSWORD_UNIT_LENGTH: 1,
 };
